Fix: Copy Link Report Enginer (#2481)
* Fix: Copy Link

* Fix: Copy Link

* Fix: Copy Link

* Fix: Copy Link

* Update report.js
