+++ src/lang/ADempiere/es/report.js
@@ -108,7 +108,9 @@ const report = {
     exportFormat: 'Formato de Exportación',
     download: 'Link de Descarga Generado',
     sharedReport: 'Reporte Compartido',
-    mesajeDownload: 'Haga clic en el Siguiente Enlace para ir a la Descarga '
+    mesajeDownload: 'Haga clic en el Siguiente Enlace para ir a la Descarga ',
+    openLink: 'Abrir Enlace',
+    mesajeShear: 'Su Enlace a Sido Copiado Exitosamente al Portapapeles'
   }
 }
