+++ src/utils/ADempiere/notification.js
@@ -80,18 +80,22 @@ export function showNotification({ type = 'success', title, message = '', summar
     }
   })
 }
-export function showNotificationReport({ type = 'success', title, message, link }) {
+export function showNotificationReport({ type = 'success', title, message, link, openLink }) {
   title = hasTranslation(title)
   if (message) {
     message = hasTranslation(message)
   }
+  openLink = hasTranslation(openLink)
   return Notification({
     title,
     message: `
-      <div style="max-height: 100px;max-width: 250px; overflow-y: auto;">
+      <div style="max-height: 100px;max-width: 250px; overflow-y: auto; text-align: center;">
         ${message}
         <br><br>
-        <a href="${link}" target="_blank" style="text-decoration: underline;">Link</a>
+        <a href="${link}" target="_blank" class="notification-link">
+          <i class="el-icon-link" style="margin-right: 5px; font-size: 16px; vertical-align: middle;"></i>
+          <span style="font-weight: bold; font-size: 14px;">${openLink}</span>
+        </a>
       </div>
     `,
     type,
